chore: some refactorings
