+++ packages/core/src/internal/verifyTransaction/rebuildTransactionPostData.ts
@@ -3,32 +3,10 @@
 
 import BigNumber from 'bignumber.js';
 import ByteBuffer from './byteBuffer';
+import {getAttachmentFields} from './getAttachmentFields';
+import {getRequestRebuildInfo} from './getRequestRebuildInfo';
+import {BaseTransaction} from './baseTransaction';
 
-interface BaseTransaction {
-    type?: number;
-    version?: number;
-    subtype?: number;
-    timestamp?: number;
-    deadline?: number;
-    senderPublicKey?: string;
-    sender?: string;
-    recipient?: string;
-    amountNQT?: string;
-    feeNQT?: string;
-    referencedTransactionFullHash?: string;
-    signature?: string;
-    flags?: number;
-    ecBlockHeight?: number;
-    ecBlockId?: string;
-    cashBackId?: string;
-}
-
-interface AttachmentField {
-    type: string;
-    parameterName?: string;
-}
-
-type AttachmentSpec = Map<string, AttachmentField[]>;
 
 /**
  * Try to rebuild the form data based on unsignedBytes in a response.
@@ -55,14 +33,11 @@ export function rebuildTransactionPostData(hexUnsignedBytes: string) {
         rebuiltData['referencedTransactionFullHash'] = transaction.referencedTransactionFullHash;
     }
 
-    const foundRequest = decodeRequestType.find(Obj => Obj.type === transaction.type && Obj.subtype === transaction.subtype);
-    if (!foundRequest) {
-        throw new Error(`Unsupported transaction type '${transaction.type}' subtype '${transaction.subtype}'.`);
-    }
-    if (foundRequest.hasAttachment) {
-        rebuiltData = parseAttachment(foundRequest.requestType, rebuiltData, trBytes);
+    const requestRebuildInfo = getRequestRebuildInfo(transaction);
+    if (requestRebuildInfo.hasAttachment) {
+        rebuiltData = parseAttachment(requestRebuildInfo.requestType, rebuiltData, trBytes);
         // Some exceptions
-        switch (foundRequest.requestType) {
+        switch (requestRebuildInfo.requestType) {
             case 'sendMoneyMultiSame':
                 rebuiltData.amountNQT = new BigNumber(rebuiltData.amountNQT)
                     .dividedBy(rebuiltData.recipients.split(';').length)
@@ -96,7 +71,7 @@ export function rebuildTransactionPostData(hexUnsignedBytes: string) {
     rebuiltData = parseEncryptToSelfMessage(transaction.flags, rebuiltData, trBytes);
 
     return {
-        requestType: foundRequest.requestType,
+        requestType: requestRebuildInfo.requestType,
         rebuiltData,
     };
 }
@@ -148,19 +123,8 @@ function parseBaseTransaction(trByteBuffer: ByteBuffer): BaseTransaction {
  * @throw Error on failure
  */
 function parseAttachment(requestType: string, data: any, trBytes: ByteBuffer) {
-    let fields: AttachmentField[];
     const attachmentVersion = trBytes.readByte();
-    switch (attachmentVersion) {
-        case 1:
-            fields = attachmentSpecV1.get(requestType);
-            break;
-        case 2:
-            fields = attachmentSpecV2.get(requestType);
-            break;
-    }
-    if (!fields) {
-        throw new Error(`Attachment specification not found for transaction type '${requestType}', version '${attachmentVersion}'.`);
-    }
+    const fields = getAttachmentFields(attachmentVersion, requestType);
     const pastValues: string[] = [];
     let sizeOfString: number;
     for (const item of fields) {
@@ -363,129 +327,3 @@ function parseEncryptToSelfMessage(transactionFlags: number, data: any, trBytes:
     return data;
 }
 
-/** From a transaction type/subtype, returns the original requestType */
-const decodeRequestType = [
-    {type: 0, subtype: 0, requestType: 'sendMoney', hasAttachment: false},
-    {type: 0, subtype: 1, requestType: 'sendMoneyMulti', hasAttachment: true},
-    {type: 0, subtype: 2, requestType: 'sendMoneyMultiSame', hasAttachment: true},
-    {type: 1, subtype: 0, requestType: 'sendMessage', hasAttachment: false},
-    {type: 1, subtype: 1, requestType: 'setAlias', hasAttachment: true},
-    {type: 1, subtype: 5, requestType: 'setAccountInfo', hasAttachment: true},
-    {type: 1, subtype: 6, requestType: 'sellAlias', hasAttachment: true},
-    {type: 1, subtype: 7, requestType: 'buyAlias', hasAttachment: true},
-    {type: 2, subtype: 0, requestType: 'issueAsset', hasAttachment: true},
-    {type: 2, subtype: 1, requestType: 'transferAsset', hasAttachment: true},
-    {type: 2, subtype: 2, requestType: 'placeAskOrder', hasAttachment: true},
-    {type: 2, subtype: 3, requestType: 'placeBidOrder', hasAttachment: true},
-    {type: 2, subtype: 4, requestType: 'cancelAskOrder', hasAttachment: true},
-    {type: 2, subtype: 5, requestType: 'cancelBidOrder', hasAttachment: true},
-    {type: 2, subtype: 6, requestType: 'mintAsset', hasAttachment: true},
-    {type: 2, subtype: 7, requestType: 'addAssetTreasuryAccount', hasAttachment: false},
-    {type: 2, subtype: 8, requestType: 'distributeToAssetHolders', hasAttachment: true},
-    {type: 2, subtype: 9, requestType: 'transferAssetMulti', hasAttachment: true},
-    {type: 20, subtype: 0, requestType: 'setRewardRecipient', hasAttachment: false},
-    {type: 20, subtype: 1, requestType: 'addCommitment', hasAttachment: true},
-    {type: 20, subtype: 2, requestType: 'removeCommitment', hasAttachment: true},
-    {type: 21, subtype: 0, requestType: 'sendMoneyEscrow', hasAttachment: true},
-    {type: 21, subtype: 1, requestType: 'escrowSign', hasAttachment: true},
-    {type: 21, subtype: 3, requestType: 'sendMoneySubscription', hasAttachment: true},
-    {type: 21, subtype: 4, requestType: 'subscriptionCancel', hasAttachment: true},
-    {type: 22, subtype: 0, requestType: 'createATProgram', hasAttachment: true},
-];
-
-const attachmentSpecV1: AttachmentSpec = new Map<string, AttachmentField[]>([
-    ['sendMoneyMulti', [
-        {type: 'Byte*1'},
-        {type: 'Long:Long*$0', parameterName: 'recipients'},
-        {type: 'Delete*1', parameterName: 'amountNQT'}
-    ]],
-    ['sendMoneyMultiSame', [
-        {type: 'Byte*1'},
-        {type: 'Long*$0', parameterName: 'recipients'},
-    ]],
-    ['setAlias', [
-        {type: 'ByteString*1', parameterName: 'aliasName'},
-        {type: 'ShortString*1', parameterName: 'aliasURI'},
-    ]],
-    ['setAccountInfo', [
-        {type: 'ByteString*1', parameterName: 'name'},
-        {type: 'ShortString*1', parameterName: 'description'},
-    ]],
-    ['sellAlias', [
-        {type: 'ByteString*1', parameterName: 'aliasName'},
-        {type: 'Long*1', parameterName: 'priceNQT'},
-    ]],
-    ['buyAlias', [
-        {type: 'ByteString*1', parameterName: 'aliasName'},
-        {type: 'Delete*1', parameterName: 'recipient'}
-    ]],
-    ['issueAsset', [
-        {type: 'ByteString*1', parameterName: 'name'},
-        {type: 'ShortString*1', parameterName: 'description'},
-        {type: 'Long*1', parameterName: 'quantityQNT'},
-        {type: 'Byte*1', parameterName: 'decimals'},
-    ]],
-    ['transferAsset', [
-        {type: 'Long*1', parameterName: 'asset'},
-        {type: 'Long*1', parameterName: 'quantityQNT'}
-    ]],
-    ['placeAskOrder', [
-        {type: 'Long*1', parameterName: 'asset'},
-        {type: 'Long*1', parameterName: 'quantityQNT'},
-        {type: 'Long*1', parameterName: 'priceNQT'}
-    ]],
-    ['placeBidOrder', [
-        {type: 'Long*1', parameterName: 'asset'},
-        {type: 'Long*1', parameterName: 'quantityQNT'},
-        {type: 'Long*1', parameterName: 'priceNQT'}
-    ]],
-    ['cancelAskOrder', [
-        {type: 'Long*1', parameterName: 'order'}
-    ]],
-    ['cancelBidOrder', [
-        {type: 'Long*1', parameterName: 'order'}
-    ]],
-    ['mintAsset', [
-        {type: 'Long*1', parameterName: 'asset'},
-        {type: 'Long*1', parameterName: 'quantityQNT'},
-    ]],
-    ['distributeToAssetHolders', [
-        {type: 'Long*1', parameterName: 'asset'},
-        {type: 'Long*1', parameterName: 'quantityMinimumQNT'},
-        {type: 'Long*1', parameterName: 'assetToDistribute'},
-        {type: 'Long*1', parameterName: 'quantityQNT'}
-    ]],
-    ['transferAssetMulti', [
-        {type: 'Byte*1'},
-        {type: 'Long:Long*$0', parameterName: 'assetIdsAndQuantities'}
-    ]],
-    ['addCommitment', [
-        {type: 'Long*1', parameterName: 'amountNQT'}
-    ]],
-    ['removeCommitment', [
-        {type: 'Long*1', parameterName: 'amountNQT'}
-    ]],
-    ['sendMoneySubscription', [
-        {type: 'Int*1', parameterName: 'frequency'}
-    ]],
-    ['subscriptionCancel', [
-        {type: 'Long*1', parameterName: 'subscription'}
-    ]],
-    ['createATProgram', [
-        {type: 'ByteString*1', parameterName: 'name'},
-        {type: 'ShortString*1', parameterName: 'description'},
-        {type: 'CreationBytes*1'},
-    ]],
-]);
-
-const attachmentSpecV2: AttachmentSpec = new Map<string, AttachmentField[]>([
-    [
-        'issueAsset', [
-            {type: 'ByteString*1', parameterName: 'name'},
-            {type: 'ShortString*1', parameterName: 'description'},
-            {type: 'Long*1', parameterName: 'quantityQNT'},
-            {type: 'Byte*1', parameterName: 'decimals'},
-            {type: 'Byte*1', parameterName: 'mintable'}
-        ]
-    ]
-]);
